linux7.0-rc1

+++ tools/testing/selftests/net/lib/py/utils.py
@@ -9,17 +9,9 @@ import subprocess
 import time
 
 
-class CmdInitFailure(Exception):
-    """ Command failed to start. Only raised by bkg(). """
-    def __init__(self, msg, cmd_obj):
-        super().__init__(msg + "\n" + repr(cmd_obj))
-        self.cmd = cmd_obj
-
-
 class CmdExitFailure(Exception):
-    """ Command failed (returned non-zero exit code). """
     def __init__(self, msg, cmd_obj):
-        super().__init__(msg + "\n" + repr(cmd_obj))
+        super().__init__(msg)
         self.cmd = cmd_obj
 
 
@@ -84,13 +76,16 @@ class cmd:
                 msg = fd_read_timeout(rfd, ksft_wait)
                 os.close(rfd)
                 if not msg:
-                    terminate = self.proc.poll() is None
-                    self._process_terminate(terminate=terminate, timeout=1)
-                    raise CmdInitFailure("Did not receive ready message", self)
+                    raise Exception("Did not receive ready message")
         if not background:
             self.process(terminate=False, fail=fail, timeout=timeout)
 
-    def _process_terminate(self, terminate, timeout):
+    def process(self, terminate=True, fail=None, timeout=5):
+        if fail is None:
+            fail = not terminate
+
+        if self.ksft_term_fd:
+            os.write(self.ksft_term_fd, b"1")
         if terminate:
             self.proc.terminate()
         stdout, stderr = self.proc.communicate(timeout)
@@ -100,21 +95,11 @@ class cmd:
         self.proc.stderr.close()
         self.ret = self.proc.returncode
 
-        return stdout, stderr
-
-    def process(self, terminate=True, fail=None, timeout=5):
-        if fail is None:
-            fail = not terminate
-
-        if self.ksft_term_fd:
-            os.write(self.ksft_term_fd, b"1")
-
-        stdout, stderr = self._process_terminate(terminate=terminate,
-                                                 timeout=timeout)
         if self.proc.returncode != 0 and fail:
             if len(stderr) > 0 and stderr[-1] == "\n":
                 stderr = stderr[:-1]
-            raise CmdExitFailure("Command failed", self)
+            raise CmdExitFailure("Command failed: %s\nSTDOUT: %s\nSTDERR: %s" %
+                                 (self.proc.args, stdout, stderr), self)
 
     def __repr__(self):
         def str_fmt(name, s):
@@ -174,11 +159,8 @@ class bkg(cmd):
         return self
 
     def __exit__(self, ex_type, ex_value, ex_tb):
-        terminate = self.terminate
-        # Force termination on exception, but only if bkg() didn't already exit
-        # since forcing termination silences failures with fail=None
-        if self.proc.poll() is None:
-            terminate = terminate or (self._exit_wait and ex_type is not None)
+        # Force termination on exception
+        terminate = self.terminate or (self._exit_wait and ex_type is not None)
         return self.process(terminate=terminate, fail=self.check_fail)
 
 
@@ -281,27 +263,9 @@ def rand_port(stype=socket.SOCK_STREAM):
     """
     Get a random unprivileged port.
     """
-    return rand_ports(1, stype)[0]
-
-
-def rand_ports(count, stype=socket.SOCK_STREAM):
-    """
-    Get a unique set of random unprivileged ports.
-    """
-    sockets = []
-    ports = []
-
-    try:
-        for _ in range(count):
-            s = socket.socket(socket.AF_INET6, stype)
-            sockets.append(s)
-            s.bind(("", 0))
-            ports.append(s.getsockname()[1])
-    finally:
-        for s in sockets:
-            s.close()
-
-    return ports
+    with socket.socket(socket.AF_INET6, stype) as s:
+        s.bind(("", 0))
+        return s.getsockname()[1]
 
 
 def wait_port_listen(port, proto="tcp", ns=None, host=None, sleep=0.005, deadline=5):
